requirements.txt
psycopg2-binary
python-dotenv
kafka-python
pyspark
pandas
